* PIMS-1233 Merged synchronization into trunk * All entities are mapped with a String UUID identifier * Fixed issues due to changing of primary key from Long to String * Added update-to-synchronization changelog(s) * Removed warehouse controller and views
git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@10991 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/order/PurchaseOrderWorkflowController.groovy
@@ -21,7 +21,7 @@ class PurchaseOrderWorkflowController {
 				flow.suppliers = orderService.getSuppliers();
 				// create a new shipment instance if we don't have one already
 				if (params.id) {
-					flow.order = Order.get(params.id as Long)
+					flow.order = Order.get(params.id)
 				}
 				else {
 					def order = new Order();
